wealth_pilot/requirements.txt
streamlit
pandas
plotly
yfinance
google-generativeai
fpdf
faker
python-dotenv
python-pptx
crewai
langchain-google-genai
